When two (or more) dambreaks map to the same flow link, the first dambreak to fail would lower the crest level. This is opposite to the general convention for weirs in which the highest weir level is used when multiple weirs are mapped to the same flow link.
To solve this we first need to determine what is the highest dambreak level and only then we can update the bob values (which contain the initial crest levels).
The new bob value will be lower than the previous bob value in case of failure, but may be overruled by non-breaching dams that are evaluated later than the breaching one, so we can't just take the minimum of old bob and new crest level.
Note that the current implementation keeps unbreached dambreak segments at the original level, but that level may be too high is it's the highest dam breaks earlier.
We should use the initialCrestLevel of the dambreaks for the unbreached crest whenever specified.